+++ src/view/manage_input/manage_post.php
@@ -1,11 +1,21 @@
 <?php 
 include_once('../../model/posts.php');
 session_start();
+if(!isset($_SESSION['uid'])){
+    echo '<script>alert("You have not logged in yet!"); 
+    window.location = "../index.php" ; </script>'; 
+}
 if (isset($_GET['post']) && ($id = $_GET['post']) ){
-    deletePost($id);
+    $curr_post = getSinglePost($id);
+    if($curr_post['uid'] == $_SESSION['uid']){
+        deletePost($id);
+    }
 }
 else if ($_POST['is_edit']){
-    updatePost($_POST['id'], $_POST['heading'], $_POST['abstract'], $_POST['content']);
+    $curr_post = getSinglePost($_POST['id']);
+    if($curr_post['uid'] == $_SESSION['uid']){
+        updatePost($_POST['id'], $_POST['heading'], $_POST['abstract'], $_POST['content']);
+    }
 }
 
 else{
